refactor: extract gateway deployment capabilities
GatewayDeploymentCapabilities was one comment away from the 999-line catalog.py cap. Move it to its own module, rewire every direct importer, and keep the package-root export. Catalog still imports the class because deployment metadata depends on it. Also add the five default-off schema-drift disclosure slots and reject grammar-tool support without custom-tool support.
